Separate model parts into more independent pieces
